requirements.txt
django
djangorestframework
django-widget-tweaks
